fix: Avoid instance label collision when creating instances
Previously, label collisions were possible because the labels were
derived from the amount of instances for the current profile. But
because it's possible to delete instances, the numbers in instance
labels aren't guaranteed to be contiguous and therefore collisions
can happen with that approach.

This patch tracks the actual highest number in instance labels and
increments that when creating new instances.